Refactor UnixShell and WindowsShell to use composition instead of inheritance
Extract all shared git/svn/phpcs workflow logic into ShellRunner, which takes
a ShellPlatform interface for platform-specific operations. Both UnixShell and
WindowsShell now implement ShellOperator and ShellPlatform directly with no
inheritance relationship between them, delegating workflow methods to ShellRunner.
